+++ SB.Boards/Queries/Stickers/PositionDto.cs
@@ -2,13 +2,13 @@
 {
     public class PositionDto
     {
-        public PositionDto(int x, int y)
+        public PositionDto(double x, double y)
         {
             X = x;
             Y = y;
         }
 
-        public int X { get; }
-        public int Y { get; }
+        public double X { get; }
+        public double Y { get; }
     }
 }
